mv bin impl to cli package
