+++ src/lib.rs
@@ -6,5 +6,7 @@ pub mod ext;
 pub mod model;
 pub mod parse;
 pub mod solve;
+pub mod cli;
 
 mod test_utils;
+
